Bug fix of `DoubleDQN`

+++ aine_drl/drl_util/decay.py
@@ -35,7 +35,7 @@ class LinearDecay(Decay):
     """
     Linear decay.
     """ 
-    def __init__(self, start_value: float, end_value: float, start_t: float, end_t: float) -> None:
+    def __init__(self, start_value: float, end_value: float, start_t: int, end_t: int) -> None:
         assert start_t < end_t and start_value >= end_value
         self.start_val = start_value
         self.end_val = end_value
